Move PIC18F4550 files to own directory
The Motor Controller could run wherever MCU (I mean
I can´t keep using too old parts that will become an
esoteric ritual to add interrupts to the project).

This commit moves it to a particular folder 'pic_controllers/
pic18f4550/' so that more support can be added to it. There
should be a tool to move this around for convenience to the
actual manufacturers IDEs, but this would be further down the
line.