Remove main C tooling and replace it with Rust toolchain
After reconsidering what I'm doing, I have decided that it would be better
to take the chance and use a different language from C. The C related files
are removed in this commit.

One of the main reasons is that I have found myself trying to find
better alternatives for unit testing since I wanted to start implementing
those, but since this is a personal project, I want to experiment in other
tools as well.

Rust has support on MCUs and the testing is embedded in the language
itself. Will give it a try.