Add the Articulation struct to represent articulations
The robot articulation encompases all points in a robot that are
relevant for the correct functioning of the robot, this includes joints,
sensors, actuators, grippers, end-effectors and any virtual articulation
that can be used to describe a position within the a robot.

So this "Articulation" is not the mere concept of a Joint, but more
refined to allow a broader interpretation of the graph that will
represent the robot.