Add KNode for the Articulation
The KNode will be the basic node that will represent a kinematic node on
a graph, however, it is the interface between the KGraph on an
SBC/MainBoard and the Articulation and Controllers themselves.

KNode get information from the other abstractions and send these to the
KGraph that has been registered. This allows not only for me to start
integrating things early, testing on actual hardware but to have the
basis for a distributed articulations within the graph network. Nice.