Update README.md
The current state of the 'README.md' doest not represent correcly what I
want to convey to anyone that ever finds the repo. The idea of to
explain that this library is meant to help describe and integrate
robotics solutions with the articulation control.

This will eventually lead to the necessity of having motion control,
path planning and many other things, which will not be possible if there
is no ground work done before.