RAC-Core: Fix KNode queue error handling
The current implementation of the KNode queues is a pointless wrapper
around a heapless BinaryHeap, this will just add complexity and it makes
the KNode harder to use. This commit removes the `rx/tx_enqueue/dequeue`
functions so that the heapless API is the one being used. Not a great
idea to reinvent the wheel.